==> metamodel_notebooks/01_download_dataset_all_columns.ipynb <==
Download raw dataset (with all columns) from www.football-data.co.uk, parse dates and save as CSV.

* Input: None
* Output: "dataset_all_columns.csv"

import pandas as pd
import numpy as np

# Downloading data.

df = pd.DataFrame()
for i in range(0, 22):
    if i < 9:
        year = f'0{i}0{i + 1}'
    elif i == 9:
        year = f'0{i}{i + 1}'
    else:
        year = f'{i}{i + 1}'
    print(f"\tDownloading season {year}")

    df_read = pd.read_csv(f'https://www.football-data.co.uk/mmz4281/{year}/SP1.csv', on_bad_lines='warn')

    try:
        df_read['Date'] = pd.to_datetime(df_read['Date'], format="%d/%m/%y")
    except ValueError:
        df_read['Date'] = pd.to_datetime(df_read['Date'], format="%d/%m/%Y")

    df = pd.concat([df, df_read])
    
df

# Saving as CSV.

df.to_csv('./data/dataset_all_columns.csv', index=False, encoding='UTF-8', sep=';')

# All teams in the dataset.

set(np.concatenate([df['HomeTeam'].unique(), df['AwayTeam'].unique()]))
