==> metamodel_notebooks/02_retrieve_bookmaker_predictions.ipynb <==
Read the full dataset and keep data of only one bookmaker and in a certain period of time.

* Input: "dataset_all_columns.csv"
* Output: "predictions_bookmaker.csv"

import pandas as pd

# Reading data.

df_read = pd.read_csv('./data/dataset_all_columns.csv', sep=';', parse_dates=['Date'], low_memory=False)
df_read

# Checking which bookmaker has more historical data.
# We are considering last 10 seasons: from 2012-08-18 to 2022-05-22.
# There are 380 matches per season, so 3800 data points would be ideal.

# It turns out that these are 100 % complete:
# B365 - Bet365
# BW - Bet&Win
# WH - William Hill

# Betting odds for weekend games are collected Friday afternoons, and on Tuesday afternoons for midweek games.

df_check = df_read.loc[ (df_read['Date']>='2012-08-18') & (df_read['Date']<='2022-05-22') ].copy()

df_check[['B365H','B365D','B365A','BSH','BSD','BSA','BWH','BWD','BWA',
          'GBH','GBD','GBA','IWH','IWD','IWA','LBH','LBD','LBA',
          'SOH','SOD','SOA','SBH','SBD','SBA','SJH','SJD','SJA',
          'SYH','SYD','SYA','VCH','VCD','VCA','WHH','WHD','WHA']].info(verbose=True, show_counts=True)

# Keeping data of only one bookmaker.

bookmaker = 'B365'

df = df_read.loc[ (df_read['Date']>='2012-08-18') & (df_read['Date']<='2022-05-22') ].copy()
df = df[['Date', 'HomeTeam', 'AwayTeam', 'FTHG', 'FTAG', bookmaker+'H', bookmaker+'D', bookmaker+'A']]

# Renaming columns.
df = df.rename(columns={
    'Date': 'date', 
    'HomeTeam': 'team_home', 
    'AwayTeam': 'team_away', 
    'FTHG': 'goals_home', 
    'FTAG': 'goals_away',
    bookmaker+'H': 'odds_home', 
    bookmaker+'D': 'odds_draw' , 
    bookmaker+'A': 'odds_away',
                    })

# Transforming odds in probabilities.
df[['odds_home', 'odds_draw', 'odds_away']] = round(1/df[['odds_home', 'odds_draw', 'odds_away']], 3)
df = df.rename(columns={
    'odds_home': 'probability_home', 
    'odds_draw': 'probability_draw' , 
    'odds_away': 'probability_away'
                    })

df

# Checking categorical predictions and observations.

predictions = []
observations = []
is_true = []

for match in df.itertuples():
    
    # Categorical prediction.
    if match.probability_home > match.probability_draw and match.probability_home > match.probability_away:
        predictions.append('home')
    elif match.probability_away > match.probability_home and match.probability_away > match.probability_draw:
        predictions.append('away')
    elif match.probability_draw > match.probability_home and match.probability_draw > match.probability_away:
        predictions.append('draw')
    else:
        predictions.append('unknown')
        
    # Categorical observation.
    if match.goals_home > match.goals_away:
        observations.append('home')
    elif match.goals_away > match.goals_home:
        observations.append('away')
    elif match.goals_home == match.goals_away:
        observations.append('draw')
    else:
        observations.append('draw')
        
    # Check if prediction is right.
    if predictions[-1] == observations[-1]:
        is_true.append(1)
    else:
        is_true.append(0)

            
df['prediction'] = predictions
df['observation'] = observations
df['is_true'] = is_true
df

# Bookmaker's accuracy.

df['is_true'].mean()

# Saving as CSV.

df.to_csv('data/predictions_bookmaker.csv', index=False, encoding='UTF-8', sep=';')
