==> metamodel_notebooks/03_join_predictions.ipynb <==
This Notebook reads the bookmaker predictions and join them with the Poisson model predictions.

* Input: "predictions_bookmaker.csv", predictions copied from "output/predictions_past" folder
* Output: "predictions_joined.csv"

import pandas as pd
import numpy as np
import glob

# Reading bookmaker predictions (one file).

df_bookmaker = pd.read_csv('./data/predictions_bookmaker.csv', sep=';', parse_dates=['date'], low_memory=False)
df_bookmaker

# Reading Poisson model predictions (folder with multiple files manually copied into "data" folder).

files_prediction_poisson = glob.glob('./data/predictions_past/predictions_train_*csv')
files_prediction_poisson.sort()

df_prediction_poisson = pd.DataFrame()
for file in files_prediction_poisson:
    df_read = pd.read_csv(file, sep=';', parse_dates=['date'], low_memory=False)
    df_prediction_poisson = pd.concat([df_prediction_poisson, df_read])
    
df_prediction_poisson

# Joining predictions together.

df = pd.merge(
    df_prediction_poisson, 
    df_bookmaker.drop(columns=['goals_home', 'goals_away', 'observation']),
    how='left', 
    on=['date', 'team_home', 'team_away'],
    suffixes=('_poisson', '_bookmaker')
)

# Deleting matches with no clear prediction.

df = df.dropna()
df = df.loc[df['prediction_bookmaker'] != 'unknown'].copy()
df = df.loc[df['prediction_bookmaker'] != 'unknown'].copy()
df

# Saving as CSV.

df.to_csv('./data/predictions_joined.csv', index=False, encoding='UTF-8', sep=';')
