==> metamodel_notebooks/04_data_enrichment.ipynb <==
Data enrichment of the joined predictions (bookmaker and Poisson) with addition information.

* Input: "predictions_joined.csv"
* Ouput: "predictions_joined_enriched.csv"

import pandas as pd
from datetime import datetime
import geopy.distance

# Reading data

df_read = pd.read_csv('./data/predictions_joined.csv', sep=';', parse_dates=['date'], low_memory=False)
df_read

# Season start and end dates

seasons_complete = [
            {
                'name': '2021-2022',
                'start_date': '2021-08-13',
                'end_date': '2022-05-22'
            },
            {
                'name': '2020-2021',
                'start_date': '2020-09-12',
                'end_date': '2021-05-23'
            },
            {
                'name': '2019-2020',
                'start_date': '2019-08-16',
                'end_date': '2020-07-19'
            },
            {
                'name': '2018-2019',
                'start_date': '2018-08-17',
                'end_date': '2019-05-19'
            },
            {
                'name': '2017-2018',
                'start_date': '2017-08-18',
                'end_date': '2018-05-20'
            },
            {
                'name': '2016-2017',
                'start_date': '2016-08-19',
                'end_date': '2017-05-21'
            },
            {
                'name': '2015-2016',
                'start_date': '2015-08-21',
                'end_date': '2016-05-15'
            },
            {
                'name': '2014-2015',
                'start_date': '2014-08-23',
                'end_date': '2015-05-23'
            },
            {
                'name': '2013-2014',
                'start_date': '2013-08-17',
                'end_date': '2014-05-18'
            },
            {
                'name': '2012-2013',
                'start_date': '2012-08-18',
                'end_date': '2013-06-01'
            },
        ]

# How far in the season are we

df = df_read.copy()
seasons_progress = []

for season in seasons_complete:
    
    df_season = df.loc[ (df['date']>=season['start_date']) & (df['date']<=season['end_date']) ].copy()
    
    start_date_season = datetime.strptime(season['start_date'], '%Y-%m-%d')
    end_date_season = datetime.strptime(season['end_date'], '%Y-%m-%d')
    length_season = end_date_season - start_date_season

    for match in df_season.itertuples():
        season_progress = (match.date - start_date_season) / length_season
        seasons_progress.append([match.date, round(season_progress, 2)])


df_seasons_progress = pd.DataFrame(data=seasons_progress, columns=['date', 'season_progress'])
df_seasons_progress = df_seasons_progress.drop_duplicates()
df_seasons_progress

df = df.merge(df_seasons_progress, left_on='date', right_on='date', how='left')
df

# Distance between stadiums
It is a measure of how far an "away team" is from home.

df_stadiums = pd.read_csv('./data/stadiums_coordinates.csv', sep=';', encoding='UTF-8')
df_stadiums.head()

distances = []

for match in df.itertuples():

    team_home = match.team_home
    team_away = match.team_away
    
    longitude_home = df_stadiums.loc[df_stadiums['team']==team_home, 'longitude'].values[0]
    latitude_home = df_stadiums.loc[df_stadiums['team']==team_home, 'latitude'].values[0]
    longitude_away = df_stadiums.loc[df_stadiums['team']==team_away, 'longitude'].values[0]
    latitude_away = df_stadiums.loc[df_stadiums['team']==team_away, 'latitude'].values[0]
    
    coordinates_home = (latitude_home, longitude_home)
    coordinates_away = (latitude_away, longitude_away)
    
    distance = geopy.distance.geodesic(coordinates_home, coordinates_away).km
    distances.append([match.date, team_home, team_away, round(distance)])

    
df_distances = pd.DataFrame(data=distances, columns=['date', 'team_home', 'team_away', 'distance'])
df_distances    

df = df.merge(df_distances, left_on=['date', 'team_home', 'team_away'], right_on=['date', 'team_home', 'team_away'], how='left')
df

# Saving as CSV

df.to_csv('./data/predictions_joined_enriched.csv', index=False, encoding='UTF-8', sep=';')
